bring multi-topic search up to par with single topic search
