+++ src/client/common/url.ts
@@ -1,6 +1,6 @@
 import qs = require('qs');
 
-export type UrlProperty = `raw` | `search` | `offset` | `limit` | `searchFrom` | `topics` | `from_time` | `to_time` | `search_by`
+export type UrlProperty = `raw` | `search` | `offset` | `limit` | `searchFrom` | `from_time` | `to_time` | `search_by`
 
 export class Url {
     obj: qs.ParsedQs
@@ -27,8 +27,12 @@ export class Url {
                 delete this.obj[prop.name]
             }
         }
+        this.Refresh()
+    }
 
-        const url = `${this.BaseUrl}?${qs.stringify(this.obj)}`
+    public Refresh() {
+        const query = qs.stringify(this.obj)
+        const url = query ? `${this.BaseUrl}?${query}` : this.BaseUrl
 
         //We're using window.history and not the router history because we don't want to navigate away, this is just for sharing url purposes.
         window.history.replaceState(null, document.title, url)
